feat: implement SSR for projects page with crawlability
- Convert projects page from client-side to server-side rendered
- Add comprehensive metadata for SEO
- Maintain interactive filtering via separate client component
- Ensure project cards are server-rendered for crawlers
- Add proper semantic HTML structure with articles
- Include all project data in initial HTML response

Resolves #249
