+++ scripts/agent-coordination.ps1
@@ -1,0 +1,135 @@
+# Multi-Agent Coordination Script
+# Provides commands and status for both agents
+
+param(
+    [Parameter(Mandatory=$true)]
+    [ValidateSet("agent1", "agent2", "status", "sync")]
+    [string]$Action
+)
+
+# Agent assignments
+$agent1Branches = @(
+    "issue-247-contact-resend-integration",
+    "issue-251-social-og-twitter-images", 
+    "issue-253-a11y-navigation-focus",
+    "issue-254-performance-images-fonts-headers"
+)
+
+$agent2Branches = @(
+    "issue-248-canonical-host-redirect",
+    "issue-249-projects-ssr-crawlability",
+    "issue-250-seo-robots-sitemap-metadata", 
+    "issue-252-remove-inflated-metrics"
+)
+
+function Show-Agent1Commands {
+    Write-Host "👤 Agent 1: Frontend/UI Specialist" -ForegroundColor Cyan
+    Write-Host "=" * 50 -ForegroundColor Cyan
+    Write-Host ""
+    Write-Host "📋 Your assigned branches:" -ForegroundColor White
+    foreach ($branch in $agent1Branches) {
+        Write-Host "  • $branch" -ForegroundColor Gray
+    }
+    Write-Host ""
+    Write-Host "🚀 Getting started:" -ForegroundColor Yellow
+    Write-Host "1. git checkout release/launch-2025-10-07" -ForegroundColor Gray
+    Write-Host "2. git pull origin release/launch-2025-10-07" -ForegroundColor Gray
+    Write-Host "3. git checkout [your-branch]" -ForegroundColor Gray
+    Write-Host ""
+    Write-Host "🔄 Daily sync:" -ForegroundColor Yellow
+    Write-Host "git checkout release/launch-2025-10-07" -ForegroundColor Gray
+    Write-Host "git pull origin release/launch-2025-10-07" -ForegroundColor Gray
+    Write-Host "git checkout [your-branch]" -ForegroundColor Gray
+    Write-Host "git merge release/launch-2025-10-07" -ForegroundColor Gray
+    Write-Host ""
+    Write-Host "📤 Create PR:" -ForegroundColor Yellow
+    Write-Host "git push origin [your-branch]" -ForegroundColor Gray
+    Write-Host "# Then create PR from your branch to release/launch-2025-10-07" -ForegroundColor Gray
+}
+
+function Show-Agent2Commands {
+    Write-Host "👤 Agent 2: Infrastructure/SEO Specialist" -ForegroundColor Magenta
+    Write-Host "=" * 50 -ForegroundColor Magenta
+    Write-Host ""
+    Write-Host "📋 Your assigned branches:" -ForegroundColor White
+    foreach ($branch in $agent2Branches) {
+        Write-Host "  • $branch" -ForegroundColor Gray
+    }
+    Write-Host ""
+    Write-Host "🚀 Getting started:" -ForegroundColor Yellow
+    Write-Host "1. git checkout release/launch-2025-10-07" -ForegroundColor Gray
+    Write-Host "2. git pull origin release/launch-2025-10-07" -ForegroundColor Gray
+    Write-Host "3. git checkout [your-branch]" -ForegroundColor Gray
+    Write-Host ""
+    Write-Host "🔄 Daily sync:" -ForegroundColor Yellow
+    Write-Host "git checkout release/launch-2025-10-07" -ForegroundColor Gray
+    Write-Host "git pull origin release/launch-2025-10-07" -ForegroundColor Gray
+    Write-Host "git checkout [your-branch]" -ForegroundColor Gray
+    Write-Host "git merge release/launch-2025-10-07" -ForegroundColor Gray
+    Write-Host ""
+    Write-Host "📤 Create PR:" -ForegroundColor Yellow
+    Write-Host "git push origin [your-branch]" -ForegroundColor Gray
+    Write-Host "# Then create PR from your branch to release/launch-2025-10-07" -ForegroundColor Gray
+}
+
+function Show-Status {
+    Write-Host "📊 Multi-Agent Development Status" -ForegroundColor Green
+    Write-Host "=" * 50 -ForegroundColor Green
+    Write-Host ""
+    
+    Write-Host "👤 Agent 1 Branches:" -ForegroundColor Cyan
+    foreach ($branch in $agent1Branches) {
+        $exists = git branch --list $branch
+        if ($exists) {
+            Write-Host "  ✅ $branch" -ForegroundColor Green
+        } else {
+            Write-Host "  ❌ $branch" -ForegroundColor Red
+        }
+    }
+    
+    Write-Host ""
+    Write-Host "👤 Agent 2 Branches:" -ForegroundColor Magenta
+    foreach ($branch in $agent2Branches) {
+        $exists = git branch --list $branch
+        if ($exists) {
+            Write-Host "  ✅ $branch" -ForegroundColor Green
+        } else {
+            Write-Host "  ❌ $branch" -ForegroundColor Red
+        }
+    }
+    
+    Write-Host ""
+    Write-Host "🔄 Current branch:" -ForegroundColor Yellow
+    $currentBranch = git branch --show-current
+    Write-Host "  $currentBranch" -ForegroundColor Gray
+}
+
+function Show-SyncCommands {
+    Write-Host "🔄 Sync Commands for Both Agents" -ForegroundColor Yellow
+    Write-Host "=" * 50 -ForegroundColor Yellow
+    Write-Host ""
+    Write-Host "📥 Pull latest changes:" -ForegroundColor White
+    Write-Host "git checkout release/launch-2025-10-07" -ForegroundColor Gray
+    Write-Host "git pull origin release/launch-2025-10-07" -ForegroundColor Gray
+    Write-Host ""
+    Write-Host "🔄 Sync your branch:" -ForegroundColor White
+    Write-Host "git checkout [your-branch]" -ForegroundColor Gray
+    Write-Host "git merge release/launch-2025-10-07" -ForegroundColor Gray
+    Write-Host ""
+    Write-Host "📤 Push changes:" -ForegroundColor White
+    Write-Host "git push origin [your-branch]" -ForegroundColor Gray
+    Write-Host ""
+    Write-Host "⚠️  Conflict Resolution:" -ForegroundColor Red
+    Write-Host "If conflicts occur, resolve them and continue:" -ForegroundColor Gray
+    Write-Host "git add ." -ForegroundColor Gray
+    Write-Host "git commit -m 'Resolve merge conflicts'" -ForegroundColor Gray
+    Write-Host "git push origin [your-branch]" -ForegroundColor Gray
+}
+
+# Main execution
+switch ($Action) {
+    "agent1" { Show-Agent1Commands }
+    "agent2" { Show-Agent2Commands }
+    "status" { Show-Status }
+    "sync" { Show-SyncCommands }
+}
